[refactor] #133 Auth Live, Test 분리
